requirements.txt
fastapi==0.118.0
uvicorn==0.37.0

SQLAlchemy==2.0.43
asyncpg==0.30.0

alembic==1.16.5
pydantic==2.11.10
pydantic_core==2.33.2

python-jose==3.5.0
passlib==1.7.4

python-dotenv==1.0.1
python-dotenv